verification/student_ss/requirements.txt
# Python dependencies to run the student TB
cocotb == 2.0.0
pyuvm == 4.0.1
pyvsc == 0.8.8.*
PyBoolector== 3.2.4.20240823.1
pytest== 9.0.2

# Visualise coverage
pyucis-viewer== 0.0.4.*
